Moved the bankruptcy model file to the root of the module
