+++ apps/legal-gazette-api/src/app/app.module.ts
@@ -19,7 +19,7 @@ import { LoggingInterceptor } from '@dmr.is/shared/interceptors'
 import { AdvertModel } from '../modules/advert/advert.model'
 import { AdvertModule } from '../modules/advert/advert.module'
 import { ApplicationModule } from '../modules/applications/application.module'
-import { BankruptcyApplicationModel } from '../modules/applications/bankruptcy/models/bankruptcy-application.model'
+import { BankruptcyApplicationModel } from '../modules/applications/bankruptcy/bankruptcy-application.model'
 import { CommonApplicationModule } from '../modules/applications/common/common-application.module'
 import { BankruptcyAdvertModel } from '../modules/bankruptcy/advert/bankruptcy-advert.model'
 import { BankruptcyDivisionAdvertModel } from '../modules/bankruptcy/division-advert/bankruptcy-division-advert.model'
