+++ apps/legal-gazette-api/src/modules/applications/application.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { SequelizeModule } from '@nestjs/sequelize'
 
 import { CaseModel } from '../case/case.model'
-import { BankruptcyApplicationModel } from './bankruptcy/models/bankruptcy-application.model'
+import { BankruptcyApplicationModel } from './bankruptcy/bankruptcy-application.model'
 import { ApplicationController } from './application.controller'
 
 @Module({
